+++ pmd-core/src/main/java/net/sourceforge/pmd/PMD.java
@@ -61,6 +61,17 @@ import net.sourceforge.pmd.util.log.ScopedLogHandlersManager;
  * process is controlled via interactions with this class. A command line
  * interface is supported, as well as a programmatic API for integrating PMD
  * with other software such as IDEs and Ant.
+ * 
+ * <p>Main entrypoints are:
+ * <ul>
+ *   <li>{@link #main(String[])} which exits the java process</li>
+ *   <li>{@link #runPmd(String...)} which returns a {@link StatusCode}</li>
+ *   <li>{@link #runPmd(PMDConfiguration)}</li>
+ *   <li>{@link #processFiles(PMDConfiguration, List, Collection, List)}</li>
+ * </ul>
+ * 
+ * <p><strong>Warning:</strong> This class is not intended to be instantiated or subclassed. It will
+ * be made final in PMD7.
  */
 public class PMD {
 
@@ -77,7 +88,11 @@ public class PMD {
 
     /**
      * Contains the configuration with which this PMD instance has been created.
+     * 
+     * @deprecated this configuration field is unused and will be removed. The class
+     * {@link PMD} should not be instantiated or subclassed.
      */
+    @Deprecated
     protected final PMDConfiguration configuration;
 
     private final SourceCodeProcessor rulesetsFileProcessor;