+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/metrics/xpath/XPathMetricFunctionTest.java
@@ -13,7 +13,7 @@ import java.util.Iterator;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import net.sourceforge.pmd.PMD;
+import net.sourceforge.pmd.PMDConfiguration;
 import net.sourceforge.pmd.PMDException;
 import net.sourceforge.pmd.Report;
 import net.sourceforge.pmd.Rule;
@@ -21,6 +21,7 @@ import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.RuleSet;
 import net.sourceforge.pmd.RuleSets;
 import net.sourceforge.pmd.RuleViolation;
+import net.sourceforge.pmd.SourceCodeProcessor;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.java.JavaLanguageModule;
 import net.sourceforge.pmd.lang.java.xpath.MetricFunction;
@@ -50,14 +51,14 @@ public class XPathMetricFunctionTest {
 
 
     private Iterator<RuleViolation> getViolations(Rule rule, String code) throws PMDException {
-        PMD p = new PMD();
         RuleContext ctx = new RuleContext();
         Report report = new Report();
         ctx.setReport(report);
         ctx.setSourceCodeFile(new File("n/a"));
         ctx.setIgnoreExceptions(false); // for test, we want immediate exceptions thrown and not collect them
         RuleSet rules = RuleSet.forSingleRule(rule);
-        p.getSourceCodeProcessor().processSourceCode(new StringReader(code), new RuleSets(rules), ctx);
+        SourceCodeProcessor sourceCodeProcessor = new SourceCodeProcessor(new PMDConfiguration());
+        sourceCodeProcessor.processSourceCode(new StringReader(code), new RuleSets(rules), ctx);
         return report.iterator();
     }